// server/api/spotify/types.ts
interface AlbumBase {
    album_type: string;
    available_markets: string[];
    copyrights: Copyright[];
    external_ids: ExternalIds;
    external_urls: ExternalUrls;
    genres: string[];
    href: string;
    id: string;
    images: Image[];
    label: string;
    name: string;
    popularity: number;
    release_date: string;
    release_date_precision: string;
    restrictions?: Restrictions;
    total_tracks: number;
    type: string;
    uri: string;
}

export interface SimplifiedAlbum extends AlbumBase {
    album_group: string;
    artists: SimplifiedArtist[];
}

export interface SavedAlbum {
    added_at: string;
    album: Album;
}

export interface Album extends AlbumBase {
    artists: Artist[];
    tracks: Page<SimplifiedTrack>;
}

export interface Albums {
    albums: Album[];
}

export interface NewReleases {
    albums: Page<SimplifiedAlbum>;
}

export interface Copyright {
    text: string;
    type: string;
}

export interface ExternalIds {
    upc: string;
}

export interface Page<TItemType> {
    href: string;
    items: TItemType[];
    limit: number;
    next: string | null;
    offset: number;
    previous: string | null;
    total: number;
}

export interface PlaylistedTrack<Item extends TrackItem = TrackItem> {
    added_at: string;
    added_by: AddedBy;
    is_local: boolean;
    primary_color: string;
    track: Item;
}

export interface AddedBy {
    external_urls: ExternalUrls;
    href: string;
    id: string;
    type: string;
    uri: string;
}

export interface LinkedFrom {
    external_urls: ExternalUrls;
    href: string;
    id: string;
    type: string;
    uri: string;
}

export interface SimplifiedTrack {
    artists: SimplifiedArtist[];
    available_markets: string[];
    disc_number: number;
    duration_ms: number;
    episode: boolean;
    explicit: boolean;
    external_urls: ExternalUrls;
    href: string;
    id: string;
    is_local: boolean;
    name: string;
    preview_url: string | null;
    track: boolean;
    track_number: number;
    type: string;
    uri: string;
    is_playable?: boolean;
    linked_from?: LinkedFrom;
    restrictions?: Restrictions;
}

export interface SavedTrack {
    added_at: string;
    track: Track;
}

export interface ExternalIds {
    isrc: string;
    ean: string;
    upc: string;
}

export interface Track extends SimplifiedTrack {
    album: SimplifiedAlbum;
    external_ids: ExternalIds;
    popularity: number;
}

export interface Tracks {
    tracks: Track[];
}

export interface SimplifiedArtist {
    external_urls: ExternalUrls;
    href: string;
    id: string;
    name: string;
    type: string;
    uri: string;
}

export interface Artist extends SimplifiedArtist {
    followers: Followers;
    genres: string[];
    images: Image[];
    popularity: number;
}

export interface Artists {
    artists: Artist[];
}

export interface FollowedArtists {
    artists: Page<Artist>;
}

export interface Followers {
    href: string | null;
    total: number;
}

export interface ExternalUrls {
    spotify: string;
}
// ...
export interface Image {
    url: string;
    height: number;
    width: number;
}
// ...
export interface Restrictions {
    reason: string;
}